The only fully built out page is: main --> laws --> education We still need to add content.
This update is essentially the completion of draft 1 of the code. Now we just add content. This edition allows one string of content per page along with a heading and subheading.

Future steps: Implement fragments to allow more diverse content. Move the content strings to SQLLite for efficient storage and retrieval.