Completion of draft 2 of the code. This edition allows one XML fragment per page along with a heading and subheading.
Commenting is sparse - feel free to add /**javadoc comments*/ liberally. It's a bit of a pain to do those all solo, and going through the code will hopefully help everybody understand it equally.

Future steps: Move the content fragments to SQLLite for efficient storage and retrieval. (Store the XML as text?)